commented out line 88 so that test 1 passes (read comments on this commit for explanations)
i'm really tired and not sure if we should remove that line OR remove the lines 52-68 that load coins which would then require the removal of line 87 (for the test to pass) and line 173 (because it becomes redundant) and then change comments on lines 163 and line 75. i'm officially too tired to make decisions so i'm going home lol. i like the loading coins code and i think it adds a lot to the tests to know that change is returned but i leave this in y'alls capable hands